fix: accept empty object, but do not drop empty workload values

+++ src/git.ts
@@ -188,14 +188,9 @@ export class Git {
   async writeFile(file: string, data: Record<string, unknown>, unsetBlankAttributes = true): Promise<void> {
     let cleanedData = data
     if (unsetBlankAttributes) cleanedData = removeBlankAttributes(data, { emptyArrays: true })
-    if (isEmpty(cleanedData)) {
-      if (file.match(secretFileRegex)) {
-        // remove empty secrets file which sops can't handle
-        return this.removeFile(file)
-      } else {
-        // Other cases of empty content will not generate valid YAML
-        throw new ValidationError('Empty payload')
-      }
+    if (isEmpty(cleanedData) && file.match(secretFileRegex)) {
+      // remove empty secrets file which sops can't handle
+      return this.removeFile(file)
     }
     // we also bail when no changes found
     const hasDiff = await this.diffFile(file, data)
